refactor: restructure consent engine to use internal package paths
