+++ exchange/consent-engine/internal/middleware/auth_test.go
@@ -11,7 +11,7 @@ import (
 	"testing"
 	"time"
 
-	"github.com/OpenNDX/openndx-core/exchange/consent-engine/v1/auth"
+	"github.com/OpenNDX/openndx-core/exchange/consent-engine/internal/auth"
 	"github.com/golang-jwt/jwt/v5"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
