+++ exchange/consent-engine/internal/router/router.go
@@ -3,9 +3,9 @@ package router
 import (
 	"net/http"
 
-	"github.com/OpenNDX/openndx-core/exchange/consent-engine/v1/auth"
-	"github.com/OpenNDX/openndx-core/exchange/consent-engine/v1/handlers"
-	"github.com/OpenNDX/openndx-core/exchange/consent-engine/v1/middleware"
+	"github.com/OpenNDX/openndx-core/exchange/consent-engine/internal/auth"
+	"github.com/OpenNDX/openndx-core/exchange/consent-engine/internal/handlers"
+	"github.com/OpenNDX/openndx-core/exchange/consent-engine/internal/middleware"
 
 	sharedUtils "github.com/OpenNDX/openndx-core/exchange/shared/utils"
 )
