+++ exchange/consent-engine/internal/services/consent_service.go
@@ -7,7 +7,7 @@ import (
 	"net/url"
 	"time"
 
-	"github.com/OpenNDX/openndx-core/exchange/consent-engine/v1/models"
+	"github.com/OpenNDX/openndx-core/exchange/consent-engine/internal/models"
 	"github.com/google/uuid"
 	"gorm.io/gorm"
 )
