+++ exchange/consent-engine/internal/utils/http_response.go
@@ -5,7 +5,7 @@ import (
 	"log/slog"
 	"net/http"
 
-	"github.com/OpenNDX/openndx-core/exchange/consent-engine/v1/models"
+	"github.com/OpenNDX/openndx-core/exchange/consent-engine/internal/models"
 )
 
 // ErrorResponse represents a standardized error response
